# Efficient labels mapping for drawing in Labels (60 FPS even with 8000x8000 images)