--- a/napari/layers/labels/_tests/test_labels.py
+++ b/napari/layers/labels/_tests/test_labels.py
@@ -18,6 +18,7 @@ from napari._tests.utils import check_layer_world_data_extent
 from napari.components import ViewerModel
 from napari.layers import Labels
 from napari.layers.labels._labels_constants import LabelsRendering
+from napari.layers.labels._labels_utils import get_contours
 from napari.utils import Colormap
 from napari.utils.colormaps import label_colormap, low_discrepancy_image
 
@@ -645,6 +646,37 @@ def test_contour_large_new_labels():
     labels_layer.contour = 1
     # This used to fail with IndexError
     viewer.dims.set_point(axis=0, value=4)
+
+
+def test_contour_local_updates():
+    """Checks if contours are rendered correctly with local updates"""
+    data = np.zeros((7, 7), dtype=np.int32)
+
+    layer = Labels(data)
+    layer.contour = 1
+    assert np.allclose(
+        layer._raw_to_displayed(layer._slice.image.raw),
+        np.zeros((7, 7), dtype=np.float32),
+    )
+
+    painting_mask = np.array(
+        [
+            [0, 0, 0, 0, 0, 0, 0],
+            [0, 0, 0, 0, 0, 0, 0],
+            [0, 0, 1, 1, 1, 0, 0],
+            [0, 0, 1, 1, 1, 0, 0],
+            [0, 0, 1, 1, 1, 0, 0],
+            [0, 0, 0, 0, 0, 0, 0],
+            [0, 0, 0, 0, 0, 0, 0],
+        ],
+        dtype=np.int32,
+    )
+
+    layer.data_setitem(np.nonzero(painting_mask), 1, refresh=True)
+
+    assert np.alltrue(
+        (layer._slice.image.view > 0) == get_contours(painting_mask, 1, 0)
+    )
 
 
 def test_selecting_label():