--- a/napari/layers/labels/labels.py
+++ b/napari/layers/labels/labels.py
@@ -899,25 +899,6 @@ class Labels(_ImageBase):
         """
         raw_modified = raw
 
-        # cache the raw data and keep track of when values are changed
-        changed_mask = None
-        if self.contour == 0:
-            if (
-                self._cached_raw_modified is not None
-                and self._cached_raw_modified.shape == raw_modified.shape
-            ):
-                changed_mask = self._cached_raw_modified != raw_modified
-                # Select only a subset with changes for further computations
-                raw_modified = raw_modified[changed_mask]
-                # Update the cache
-                self._cached_raw_modified[changed_mask] = raw_modified
-            else:
-                self._cached_raw_modified = raw_modified.copy()
-
-            # If there are no any changes, just return the cached image
-            if raw_modified.size == 0:
-                return self._cached_image
-
         if self.contour > 0:
             if raw.ndim == 2:
                 raw_modified = np.zeros_like(raw)
@@ -937,6 +918,25 @@ class Labels(_ImageBase):
                         deferred=True,
                     )
                 )
+
+        # cache the raw data and keep track of when values are changed
+        changed_mask = None
+        if (
+            self._cached_raw_modified is not None
+            and self._cached_raw_modified.shape == raw_modified.shape
+        ):
+            changed_mask = self._cached_raw_modified != raw_modified
+            # Select only a subset with changes for further computations
+            raw_modified = raw_modified[changed_mask]
+            # Update the cache
+            self._cached_raw_modified[changed_mask] = raw_modified
+        else:
+            self._cached_raw_modified = raw_modified.copy()
+
+        # If there are no any changes, just return the cached image
+        if raw_modified.size == 0:
+            return self._cached_image
+
         if self._color_lookup_func is None:
             self._color_lookup_func = self._get_color_lookup_func(
                 raw_modified, np.min(raw_modified), np.max(raw_modified)