--- a/napari/layers/labels/labels.py
+++ b/napari/layers/labels/labels.py
@@ -901,21 +901,22 @@ class Labels(_ImageBase):
 
         # cache the raw data and keep track of when values are changed
         changed_mask = None
-        if (
-            self._cached_raw_modified is not None
-            and self._cached_raw_modified.shape == raw_modified.shape
-        ):
-            changed_mask = self._cached_raw_modified != raw_modified
-            # Select only a subset with changes for further computations
-            raw_modified = raw_modified[changed_mask]
-            # Update the cache
-            self._cached_raw_modified[changed_mask] = raw_modified
-        else:
-            self._cached_raw_modified = raw_modified.copy()
+        if self.contour == 0:
+            if (
+                self._cached_raw_modified is not None
+                and self._cached_raw_modified.shape == raw_modified.shape
+            ):
+                changed_mask = self._cached_raw_modified != raw_modified
+                # Select only a subset with changes for further computations
+                raw_modified = raw_modified[changed_mask]
+                # Update the cache
+                self._cached_raw_modified[changed_mask] = raw_modified
+            else:
+                self._cached_raw_modified = raw_modified.copy()
 
-        # If there are no any changes, just return the cached image
-        if raw_modified.size == 0:
-            return self._cached_image
+            # If there are no any changes, just return the cached image
+            if raw_modified.size == 0:
+                return self._cached_image
 
         if self.contour > 0:
             if raw.ndim == 2: