… napari/benchmarks/benchmark_labels_layer.py
… napari/utils/colormaps/colormap_utils.py