# Description

This PR introduces two independent optimizations for labels rendering during drawing: local updates and caching optimizations.

These optimizations virtually solve any performance issues related to drawing in the Labels layer. With them, it is possible to achieve 60 FPS when drawing even in 8000x8000 images.

### Local updates

The initial implementation updates the whole labels map and copies it to the shader on each brush update, which becomes a major bottleneck when you work with high-resolution images no matter how effective the labels mapping implementation is. 

In this PR, only partial updates are sent to the VisPy labels layer. Each time `Layers.data_setitem` is called, it tracks the bounds of modified region, and instead of calling `Labels.refresh()` that triggers the update of the whole labels image, it calls `Labels._partial_labels_refresh` that emits the `Labels.events.labels_update` event that comes with the slice localizing the modified region, which is then handled by `VispyLabelsLayer._on_partial_labels_update`. VisPy textures can be partially updated, which is used in the `_on_partial_labels_update` method.

### Caching optimization

The idea is to recompute color mapping only for the elements of a label map that are changed from a previous update. Even in the parts of the code where the color mapping is implemented quite efficiently it can give up to 5x speedup by avoiding slow `np.float32` recomputations for most pixels.

In practice, when you use brush, less than 1% of pixels are updated at each iteration (on large images the percentage is even smaller). As a result, this optimization should work and give a significant boost in 99%+ of typical use case scenarios.

## Benchmark results

I added a new benchmark (benchmark_labels_layer.LabelsDrawing2DSuite) that measures the timings of brush drawing in different modes (auto/direct, contour == 0/1) with different brush sizes. It simulates the brush drawing from the position (0, 0) to (n - 1, n - 1) with 30 refresh updates along the way.

This PR:
```                                                                                                                          
              ====== ============ ============ ============ ============ ============
              --                                  color_mode / contour
              ------------------- ---------------------------------------------------
                n     brush_size    auto / 0     auto / 1    direct / 0   direct / 1
              ====== ============ ============ ============ ============ ============
               512        8        28.2±0.3ms    33.6±1ms     29.2±1ms    32.9±0.7ms
               512        64       22.8±0.1ms   28.1±0.2ms   22.9±0.1ms   29.1±0.3ms
               512       256       128±0.9ms    153±0.7ms    128±0.7ms     153±1ms
               3072       8        147±0.6ms    165±0.7ms     147±3ms     166±0.4ms
               3072       64       131±0.9ms     151±2ms     132±0.8ms     150±1ms
               3072      256        450±2ms      502±2ms     452±0.7ms     501±2ms
              ====== ============ ============ ============ ============ ============
```

main 7b3f7ecfe0e74af5f52dd3eb380b5752b2f3378c:
```
              ====== ============ ============ ============ ============ ============
              --                                  color_mode / contour
              ------------------- ---------------------------------------------------
                n     brush_size    auto / 0     auto / 1    direct / 0   direct / 1
              ====== ============ ============ ============ ============ ============
               512        8        86.6±0.7ms    242±2ms      100±1ms      256±1ms
               512        64        75.6±1ms     230±1ms      90.6±1ms     241±1ms
               512       256       163±0.9ms     316±2ms      180±1ms      328±1ms
               3072       8        1.20±0.01s   6.33±0.02s   1.80±0.02s   6.90±0.02s
               3072       64       1.15±0.01s    6.30±0s      1.77±0s      6.90±0s
               3072      256       1.30±0.01s   6.40±0.01s    1.92±0s     7.04±0.02s
              ====== ============ ============ ============ ============ ============
```

The difference is most noticeable when the heavy labels mapping code is used (contour=1), with small brush sizes it gives almost 50x performance boost.

Please keep in mind that this benchmark only measures the time it takes to update `layer.data` + the time it takes to convert labels to colors before sending them to VisPy. This benchmark does not take into account the time it takes to transfer data from napari to OpenGL or the time it takes to process labels in any type of OpenGL shader. And the local updates also give a substantial speed up in the part that is not measured.

## Type of change
<!-- Please delete options that are not relevant. -->
- [X] Optimization (non-breaking change which speedups existing code)

# How has this been tested?
- [X] all tests pass with my change

## Final checklist:
- [X] My PR is the minimum possible work for the desired functionality
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have made corresponding changes to the documentation
- [X] I have added tests that prove my fix is effective or that my feature works
- [ ] If I included new strings, I have used `trans.` to make them localizable.
      For more information see our [translations guide](https://napari.org/developers/translations.html).