@ksofiyuk 2023-04-15
@alisterburt Actually, there is a small problem. It failed some tests for the Label layer. It turned out that this optimization cannot be easily applied when `Layer.contour > 0` as it performs morphological operations on the `raw` array. I'll turn it off.