@ksofiyuk 2023-04-15
I found out that there is no need to disable it when `contour > 0`, it is enough just to move the cache initialization after the if block, in which contours are computed. However, computing contours at each update makes caching much less useful, but it should give some gain anyway.