@ksofiyuk 2023-04-17
> Another reason why I want to have the two reviews separately is that I think there should be a bit more of a change in the middle to make it clearer what's going on.

@jni No problem at all, I assumed to have them reviewed separately that's why I created a separate pull request. 

>  For example, I don't like reusing variable names (raw_modified = raw_modified[changed_mask]), as it makes it harder to reason about the code (is this image shaped or a linear array of changed values?). I also don't like that now, in most cases, image =  will actually not be an image, but a linear set of values. So I'd like a bit of an update to the logic and variable names to better reflect what's going on.

I agree on the part on the `image` naming, it's confusing. Maybe it's better to use something like `mapped_labels` instead? This name doesn't have any assumptions about the dimensionality of an array and I think it describes the variable better.

If we want to refactor the variables names in this method, then I suggest changing `raw` to `labels` as I think it describes the object better. `raw_modified` to `labels_modified` or `labels_changed`. I don't fully agree about a remark on reusing `raw_modified` as this PR actually makes this initial name more appropriate in this method. As you can interpret it as a part of `raw` that is modified and needs to be recalculated. In the initial implementation the name `raw_modified` was confusing for me. Why was it called `modified` if it was equal to `raw` (most common scenario), except for the case when contours are used?

> I thought that this would actually track changed pixels from painting and so on, in which case the speedup would be obvious. But no, it actually checks for changes by doing a full array != value, which is not immediately obviously faster than doing array.astype(). But your benchmarks (and my own experimentation with this PR pray) suggest that it definitely is faster! Do I understand correctly that that's what you're doing here?

You understood everything absolutely correctly. I implemented this caching only to avoid recomputation of the label mapping in floats. So it replaces several slow float operations on a full array with just a single comparison of integers and taking a subarray using a boolean mask. That's where the performance boost comes from.

> I thought that this would actually track changed pixels from painting and so on, in which case the speedup would be obvious

I'm still thinking that it is a good thing to do, but it requires a deep refactoring not only of the Label layer, but also changing the way how the labels are sent to GPU. In the current implementation, optimizing `_raw_to_displayed` further just doesn't make any sense as the whole array is sent to VisPy which performs a lot of float operations with this array before even sending it to GPU, which is now a bottleneck.

I think it's better to do it based on #3308.

> could you please test this solution with zarr, tensorstore, or dask arrays as labels? I'm concerned about some of these operations (boolean indexing) in those contexts...

I assume they should follow the numpy indexing convention, otherwise there are a lot of places where they can be broken. For example, in the contours computation code:
https://github.com/napari/napari/blob/5e8ffc6c1bd5d8dcd3b507b64a84cb4c5d1a2c1d/napari/layers/labels/labels.py#L905-L913

Btw, does napari have any examples of using these arrays as labels?