@Czaki 2023-04-17
I think that they think about resolve merge conflicts.